requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.22
langchain-openai==1.1.12
langchain-postgres==0.0.17
langchain-text-splitters==1.1.1
langsmith==0.7.22
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.3
openai==2.30.0
orjson==3.11.7
packaging==26.0
pgvector==0.3.6
propcache==0.4.1
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pypdf==6.9.2
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.48
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
